masters/requirements.txt
Django==2.1.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
pytz==2020.1
